repair: extract row_level_repair
