+++ configure.py
@@ -861,6 +861,7 @@ scylla_core = (['database.cc',
                 'repair/writer.cc',
                 'repair/repair.cc',
                 'repair/row_level.cc',
+                'repair/row_level_repair.cc',
                 'exceptions/exceptions.cc',
                 'auth/allow_all_authenticator.cc',
                 'auth/allow_all_authorizer.cc',
